Adding `MANIFEST.in` to the project.
Right now it includes:

- the docs/ directory
- the examples/ directory
- the test/ directory
- the Makefile
- the CHANGES file
- the LICENSE file
- the AUTHORS file

It also provides exlusion for the docs/_build directory which generates the
documentation files (from sphinx).

Some of the items it includes are not yet finished, or even created at all.
This is simply being added to the project to prepare for these items to be
added.