Adding a few more doctests.
These tests test very very basic exception handling for the `spool` class
method of the `CallFile` class.

The first test makes sure that either the `channel` or `trunk_type`,
`trunk_name`, and `number` attributes are specified, since Asterisk requires a
'Channel:' directive to be present in every call file.

The second test makes sure that either the `application` or `extension`
attributes are specified because Asterisk requires one of them to be specified
as well.