Adding setup.cfg to the project.
This file contains a configuration directive for `nose` to be used with
setuptools. In particular, adding this file to the project lets us do:

python setup.py nosetests

To run all of our tests. This is important because by running our tests in this
manner, we are able to use setuptools to auto build our package and check for
errors before each test run, and also install any package dependencies before
running tests as well.

This makes testing convenient and portable.