Removing `trunk_type`, `trunk_name`, and `number` attributes.
These attributes were previously used as alternate (required) attributes.

Either a user had to specify: channel, or trunk_type, trunk_name, and number.

The reason I'm removing them is because giving developers a choice of which one
to supply is rather sinister in a hidden way:

The Asterisk call file `Channel:` attribute accepts a Dial string, eg:

	Technology/Resource[&Technology2/Resource2[&...]

So potentially, users can specify multiple numbers to call with multiple
technologies, not just one. This is deceptive to developers because if they
specify the channel variable, they can make full use of the Channel attribute,
but if they use trunk_type, trunk_name, and number, they are limited to a
single number.

So instead of making some clusterfuck of dictionaries of numbers or something
(ok, well maybe a future version will make it easier), I'm removing those three
attributes so that developers must specify the dial string themselves. At least
this way it won't be deceptive.

As an added benefit, this also shrinks the code paths in pycall, and will make
testing a bit easier.