Changing docstring for `filename` property.
Instead of returning the absolute path of our call file. It will return the
basename (without the directory part).

Why?

Because, we make the file in two places--we create it ONCE as a temporary file,
then we MOVE it to the spoolding directory. It MUST be done in two steps
because the asterisk spooling daemon is aggressive, and will try to execute a
half-written call file. This way, the `filename` property won't return an
ambiguous filename for the callfile.