Removing `tmpdir` and `_filename` attributes.
These attributes previously allowed developers to override some temporary call
file information.

`tmpdir` would allow devs to specify the temporary directory where call files
are created before they are moved to the spooling directory.

`_filename` allowed devs to change the actual name of the call file.

I'm removing both of these because I think they are both unnecessary, and
cumbersome.

`tmpdir` is not necessary because it doesn't matter where the temporary file is
created. python's `mkstemp` function writes the file securely to a temporary
directory, so overriding this behavior is really just a waste of time. Since
the temporary files only exist for a fraction of a second anyways, it's
pointless.

`_filename` is equally pointless. The main goal of this library is to generate
call files quickly and easily. Manually specifying a filename just doesn't make
sense. It can also cause unexpected problems if the developer clobbers call
files that he intended to store.

These changes will drastically reudce the code size and simplify tests a lot. I
think that simplicity > hack-a-bility in this case.