Fixing bug that effect all int arguments.
This includes `wait_time`, `retry_time`, and `max_retries`. The bug was that if
a user set a value of 0 for any of those, we wouldn't write that value to the
call file. This is a mistake, because python evaluates 0 AND None to False.

To fix it, I simply added an explicit check for `None`, which looks ugly, but
works.