Removing the `filename` property.
This was added at the request of another user. I think it is unnecessary, so
I'm removing it.

This library is really meant to write call files quickly and in bulk, so it
doesn't make sense to allow retrieval of a filename. If anything, we should
return the written filename after we write the call file, and that's it.