Adding test case for spool method.
This test ensures that the spool method raises InvalidTimeError if the user
specifies as invalite time parameter to the spool method.

This could be really helpful, especially if the developer is new to pycall, and
doesn't realize that the time parameter is only valid if it is a datetime
object.

This could save a lot of frustration and debugging for developers later on.